Add linter for Solidity files (#436)
* lint solidity example files

* lint solidity in /contracts/src

* extract mock imageid script

* use ierc20 instead of custom interface or just address

* disable specific warnings

+++ examples/erc20_balances/src/vlayer/VToyken.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import {IERC20} from "openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
 import {ERC20} from "openzeppelin-contracts/contracts/token/ERC20/ERC20.sol";
 
 contract VToyken is ERC20 {
@@ -18,7 +17,7 @@ contract VToyken is ERC20 {
             0xa0Ee7A142d267C1f36714E4a8F75612F20a79720
         ];
 
-        for (uint i = 0; i < initialOwners.length; i++){
+        for (uint256 i = 0; i < initialOwners.length; i++){
             _mint(initialOwners[i], i * 1 ether);
         }
     }
