+++ examples/simple_travel/src/vlayer/SimpleTravelProver.sol
@@ -11,8 +11,8 @@ contract SimpleTravelProver is Prover {
     function aroundTheWorld() public returns (uint256) {
         // Important: the address of otherChainContract depends on when it was deployed on anvil 2.
         address otherChainContract = 0x5FbDB2315678afecb367f032d93F642f64180aa3;
-        uint testChainId2 = 1114;
-        uint blockNo = 1;
+        uint256 testChainId2 = 1114;
+        uint256 blockNo = 1;
 
         setChain(testChainId2, blockNo);
         return IExample(otherChainContract).example();
